MCPP: fixed 'git apply' issues on Windows
MCPP was failing to compile because it was using 'git apply'
to apply the required patch that contains the O3DE code changes.

'git apply' fails silently when it runs outside of a git repository.
To avoid this problem 'get_and_build_mcpp.py' now creates a local
git repository before calling 'git apply patch'

It was found that 'git apply' would fail in Windows because of the
LF to CRLF auto conversion. To circumvent this issue , added
.gitattributes file to force the patch file to be fetched without CRLF,
keeping its original LF characters. Also the local git repo is created
with ['git', 'config', '--local', 'core.autocrlf', 'false'] and
the patching works successfully.

Additionally added checking for the x64 version of the Visual Studio
Command Prompt which by default is x86. This will force the user
to manually run the x64 version. The x86 version will compile but it
will be incompatible with O3DE.

Signed-off-by: garrieta <garrieta@amazon.com>