MCPP: Clean Up First Patch From GPL Violations
Removed the original 'mcpp-2.7.2-az.2.patch' as it had unnecessary
changes to several config/* files that had GPL licenses.

The new, clean, patch 'mcpp_2.7.2_az.patch' doesn't have '.2.' as there
will always be one patch file, that may change in the future and it will
be easier to see the differences in future code reviews.

Signed-off-by: garrieta <garrieta@amazon.com>