OpenEXR fix min osx version, bugfixes to scripts
The OpenEXR package was not setting the correct "Minimum OS Version"
on mac, and thus when linked, would produce "This is for a later
operating system version than the project you're building is for"
warnings in O3DE.

While fixing this, it was found that pull_and_build_from_git was
cleaning its install folder AFTER already copying some files there
like the license file.  This caused invalid builds but only when
filtering was not used.

While fixing this it was also discovered that the method for
importing the configs from the OpenEXR upstream package would be
sticky - that is, it would still attach to the old version of
the package even if you got a new one due to the way find_package
works with CONFIG mode, and how it caches the _DIR variable.

To fix this, a chunk of code was added to clear out those cache
variables before calling find_package.  Becuase this code needed
to be in multiple places, and so does the import mapping, both of those
functions were moved to a common file.

Becuase this common file could be used by OTHER packages, the common
file was added to the repo scriptsin general so anyone can use it.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>