Android openssl 1.1.1o fix - uses older NDK now
The previous version of the package used a later version of the
NDK than the rest of the engine did, causing compatibility issues
with build farms that used older NDK.

As part of this change, it also specifies the min Android system
API level (30) so as not to use "latest", as well as prints out
the NDK location as part of the build.

This ONLY revs the android package, as it was the only one affected.

Q:  How do we "prove" that this package will not cause the same problem
A:  We use the android native tools version of the 'nm' tool, which
dumps symbols.

The old package:
```
aarch64-linux-android-nm libcrypto.a  | grep -i ldadd4_relax
                 U __aarch64_ldadd4_relax
                 U __aarch64_ldadd4_relax
                 ... (many references)
                 U __aarch64_ldadd4_relax
                 U __aarch64_ldadd4_relax
```

The new package:
```
aarch64-linux-android-nm libcrypto.a  | grep -i ldadd4_relax
(no output)
```

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>