Adding package scripts to build the sdformat13 library
The scripts build the sdformat13 library and its dependent libraries of tinyxml2, gz-cmake, gz-utils2, gz-math7 as well.

The build scripts are available to be run using either an Ubuntu 20.04+ distro or the ArchLinux distro.

The actual building of the library takes place in a Ubuntu 20.04 Docker image and the install artifacts are copied out of it.

The available libraries for use are exposed via the `3rdParty::sdformat` target.

The binary artifacts are `libsdformat13.so, `libtinyxml.a`, `libgz-utils2.so` and `libgz-math7.so`.

The package has been verified to build for both `x86_64` and `aarch64` Linux architectures and have been tested via adding using it as a dependency of the ROS2 Gem via specifying a local package directory via the `LY_PACKAGE_SERVER_URLS` ENV.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>