Fixed building of libsdformat through the 3p
The cmake configure command for libsdformat contained inadvertent [added text](https://github.com/o3de/3p-package-source/commit/4e3d9ba37c1dde699febc4f64942d2b08eece587#diff-f7b9acab9d6b21567f4c1bd73f7e353a1bab0f3d418dff40569ab812928386dcR193) which was causing it to fail to configure.

Updated libsdformat to build using the configuration of Release with Debug Info.
The Debug Info information is then detached from the libsdformat DSO and kept as a sidecar .dbg file next to it, that will load the debug information if available.

This allows the the sdformat data structures to be seen in the debugger if the `libsdformat13.so.13.5.0.dbg` file has not been moved or deleted, but also doesn't increase the size of the actual `libsdformat13.so.13.5.0` library itself.

This also provides a mechanism for linux/android 3rdParty libraries to be built in release and have debug information available for users of the engine via source. But also makes it possible for the SDK build or installer to skip copying over the `.dbg` files with debug information and therefore allow the SDK to be smaller.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>