Update DXC to version 2505.1-o3de (#289)
Update DirectXShaderCompiler to the current version.

The new version is based on the 2505.1 release of DXC with some
additional commits from o3de. The additional commits are the same as in
the last version.

We want this update because we run into a compiler bug in the 2308
version. Some shader produced the error message `DXC: Int: 2` or `DXC:
Int: 3`. We're not sure where this is coming from, but it is fixed in
the newer version.

I tested this update with the MainRenderPipeline and a custom render
pipeline under Windows with both DX12 and Vulkan.

This PR also fixes two problems in the Github Actions:
- Update Windows to latest and use VS2022 instead of 2019 for build DXC.
windows-19 does not work anymore in Github Actions
- Don't set the cmake environment variables to use CCache when CCache is
not used

---------

Signed-off-by: Martin Sattlecker <martin.sattlecker@huawei.com>